Prevenção de erro no aws-s3

+++ src/server/routes/controllers/brand.js
@@ -169,6 +169,10 @@ exports.remove = (req, res) => {
           }
 
           const delFilesFromS3 = (next) => {
+            if (!ObjectsS3ForDelete[0]) {
+              return next && next()
+            }
+
             s3
               .deleteObjects({
                 Bucket: process.env.BUCKET_NAME,
